requirements.txt
zhipuai
PyYAML